wip: add global prefix
It does not work yet. For the moment it is also based on the fact that
we strip the prefix in traefik before passing it to the wikijs.
For the moment, the only thing that works is the rendering of the pages:
all the internal links have prefix added. The nice thing is that the
prefix can be changed but just re-rendering all the pages.